+++ Pandora/Paradigm/Structure/Concrete.hs
@@ -1,3 +1,4 @@
 module Pandora.Paradigm.Structure.Concrete (module Exports) where
 
+import Pandora.Paradigm.Structure.Concrete.Graph as Exports
 import Pandora.Paradigm.Structure.Concrete.Stack as Exports