+++ CHANGELOG.md
@@ -10,5 +10,6 @@
 
 # 0.1.2
 * Define `Pipeline` control flow paradigm
-* Split `Strucure` modules on `Property` and `Concrete`
+* Split `Structure` modules on `Property` and `Concrete`
 * Define `Hollow` type class for handling empty structures
+* Define `Graph` concrete structure