+++ CHANGELOG.md
@@ -16,3 +16,6 @@
 * Define `Graph` concrete structure
 * Define infix `:-.` type operator for Lens
 * Define `Object` instances for `Product` datatype
+
+# 0.1.2
+* Define `Semigroup` and `Monoid` instances for `Cofree` datatype